revert last commit, I'm too nervous

+++ pioreactor/utils/timing.py
@@ -46,7 +46,7 @@ class RepeatedTimer:
         the job name that is called RepeatedTimer - will be included in logs
     run_immediately: bool
         The default behaviour is to wait `interval` seconds, and then run. Change this to True to run `func` immediately.
-    run_after: int
+    run_after: float
         After calling `start`, wait for `run_after` seconds, then continue as normal. This happens before `run_immediately`.
     args, kwargs:
         additional arg and kwargs to be passed into function.
